Dataserver/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
groq>=1.2.0
reportlab>=4.0
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
pydantic>=2.7
requests==2.32.3
